moved resolvers and tests to submission component
