+++ packages/component-submission/server/resolvers/index.test.js
@@ -40,3 +40,34 @@ describe('Manuscripts', () => {
     })
   })
 })
+
+describe('component-submission resolvers', () => {
+  beforeEach(async () => {
+    await createTables(true)
+    replaySetup('success')
+  })
+
+  describe('editors', () => {
+    it('returns a list of senior editors', async () => {
+      const result = await Query.editors({}, { role: 'senior-editor' })
+      expect(result.length).toBeGreaterThanOrEqual(40)
+      expect(result[0]).toEqual({
+        id: '8d7e57b3',
+        aff: undefined,
+        name: 'Richard Aldrich',
+        focuses: [
+          'ion channels',
+          'calcium binding proteins',
+          'membrane transport',
+          'allostery and cooperativity',
+          'cellular neurophysiology',
+          'biochemical neuroscience',
+        ],
+        expertises: [
+          'Structural Biology and Molecular Biophysics',
+          'Neuroscience',
+        ],
+      })
+    })
+  })
+})
