+++ packages/component-submission/server/typeDefs.graphqls
@@ -1,3 +1,7 @@
+extend type Query {
+  editors(role: String!): [EditorAlias]
+}
+
 extend type Mutation {
   uploadManuscript(id: ID!, file: Upload!, fileSize: Int!): Manuscript!
   updateManuscript(data: ManuscriptInput!): Manuscript!
