+++ server/xpub-server/entities/user/resolvers.js
@@ -1,13 +1,7 @@
 const jwt = require('jsonwebtoken')
 const config = require('config')
-const elifeApi = require('@elifesciences/component-model-user/entities/user/helpers/elife-api')
 
 const resolvers = {
-  Query: {
-    async editors(_, { role }) {
-      return elifeApi.people(role)
-    },
-  },
   Mutation: {
     async exchangeJournalToken(_, { token }) {
       const secret = config.get('pubsweet-server.secret')
